Moved to float32 system instead

+++ tools/Color.go
@@ -2,5 +2,5 @@ package tools
 
 type Color struct {
 	// TODO: Add reflection, roughness, and diffusion
-	R, G, B, A float64
+	R, G, B, A float32
 }
